Fix setupIonicSvelte import
Uses await import() onMount. Assumption is there must be something that is executing server-side, and I think svelte-kit still works as if it's in SSR mode whilst in dev. Importing onMount ensures this code is only read client-side.

One side-effect from the above, is that there is FOUC until setupIonicSvelte is executed. A small fix for this is to simply set a `ionicLoaded` variable, to check against, to block rendering until ionic is ready. It isn't ideal, but I think it may be the best we can do.